msm: kgsl: sysfs store and show should return ssize_t
Some of these functions were declared to return int,
which isn't correct.

+++ drivers/gpu/msm/adreno.c
@@ -2274,12 +2274,12 @@ static struct adreno_device *_get_adreno_dev(struct device *dev)
  * KGSL_FT_SKIPFRAME -> BIT(3) Set to skip frame
  * by default set FT policy to KGSL_FT_DEFAULT_POLICY
  */
-static int _ft_policy_store(struct device *dev,
+static ssize_t _ft_policy_store(struct device *dev,
 				     struct device_attribute *attr,
 				     const char *buf, size_t count)
 {
 	struct adreno_device *adreno_dev = _get_adreno_dev(dev);
-	int ret;
+	ssize_t ret;
 	if (adreno_dev == NULL)
 		return 0;
 
@@ -2298,7 +2298,7 @@ static int _ft_policy_store(struct device *dev,
  *
  * This is a routine to read current FT policy
  */
-static int _ft_policy_show(struct device *dev,
+static ssize_t _ft_policy_show(struct device *dev,
 					struct device_attribute *attr,
 					char *buf)
 {
@@ -2325,7 +2325,7 @@ static int _ft_policy_show(struct device *dev,
  * KGSL_FT_PAGEFAULT_LOG_ONE_PER_INT -> BIT(3) Set to log only one
  * pagefault per INT.
  */
-static int _ft_pagefault_policy_store(struct device *dev,
+static ssize_t _ft_pagefault_policy_store(struct device *dev,
 				     struct device_attribute *attr,
 				     const char *buf, size_t count)
 {
@@ -2364,7 +2364,7 @@ out:
  *
  * This is a routine to read current FT pagefault policy
  */
-static int _ft_pagefault_policy_show(struct device *dev,
+static ssize_t _ft_pagefault_policy_show(struct device *dev,
 					struct device_attribute *attr,
 					char *buf)
 {
@@ -2385,7 +2385,7 @@ static int _ft_pagefault_policy_show(struct device *dev,
  * 0x1 - Enable fast hang detection
  * 0x0 - Disable fast hang detection
  */
-static int _ft_fast_hang_detect_store(struct device *dev,
+static ssize_t _ft_fast_hang_detect_store(struct device *dev,
 				     struct device_attribute *attr,
 				     const char *buf, size_t count)
 {
@@ -2429,7 +2429,7 @@ static int _ft_fast_hang_detect_store(struct device *dev,
  * @attr: Device attribute
  * @buf: value read
  */
-static int _ft_fast_hang_detect_show(struct device *dev,
+static ssize_t _ft_fast_hang_detect_show(struct device *dev,
 					struct device_attribute *attr,
 					char *buf)
 {
@@ -2451,7 +2451,7 @@ static int _ft_fast_hang_detect_show(struct device *dev,
  * 0x0 - Enable long IB detection
  * 0x1 - Disable long IB detection
  */
-static int _ft_long_ib_detect_store(struct device *dev,
+static ssize_t _ft_long_ib_detect_store(struct device *dev,
 				     struct device_attribute *attr,
 				     const char *buf, size_t count)
 {
@@ -2475,7 +2475,7 @@ static int _ft_long_ib_detect_store(struct device *dev,
  * @attr: Device attribute
  * @buf: value read
  */
-static int _ft_long_ib_detect_show(struct device *dev,
+static ssize_t _ft_long_ib_detect_show(struct device *dev,
 					struct device_attribute *attr,
 					char *buf)
 {
